Update to fix issue comparing 2 signed and unsigned number

+++ solutions/code_forces/a/perfect_permutation.cc
@@ -11,7 +11,7 @@ using namespace std;
 */
 int main() {
   //init
-  int input;
+  size_t input;
   vector<int> ret;
 
   //read input
